Convert lightLevel to 0-15 instead 0.0-1.0

+++ src/main/java/me/otho/customItems/configuration/jsonReaders/blocks/Cfg_block.java
@@ -17,7 +17,7 @@ public class Cfg_block extends Cfg_basicData {
   public String toolClass;
   public float resistance = 10;
   public float hardness = 2;
-  public float lightLevel = 0;
+  public int lightLevel = 0;
   public int harvestLevel = 0;
   public float slipperiness = 0.6f;
   public boolean isCollidable = true;
